+++ stlexporter/stl_exporter.py
@@ -2,7 +2,8 @@ from glue.config import viewer_tool
 from glue.viewers.common.tool import Tool
 from stlexporter import ICON
 
-from .ui import MainWindow
+from stlexporter.ui import MainWindow
+from stlexporter.controller import Controller
 
 @viewer_tool
 class StlExporter(Tool):
@@ -15,13 +16,12 @@ class StlExporter(Tool):
         super(StlExporter, self).__init__(viewer)
 
     def activate(self):
-        dialog = MainWindow()
         viewer = self.viewer
-
-        #get the data layers of the viewer.
         layers =  viewer.state.layers
+        
+        dialog = MainWindow(Controller(viewer, layers))
 
-        dialog.create_layout(viewer, layers)
+        dialog.createMainWindow()
         dialog.show()
     
     def close(self):
